examples/custom-python/requirements.txt
-e ../../packages/python/custom
litellm
python-dotenv
